:x: refactor for clarity, better error states
