+++ app/views/location.nunjucks
@@ -22,7 +22,7 @@
       {% else %}
         <div class="form-group">
       {% endif %}
-        <label class="form-label-bold" for="location">{{ locationLabel }}</label>
+        <label class="form-label-bold" for="location">Enter a postcode.</label>
           <div class="form-group">
             <input autocomplete="off" type="text" class="form-control" name="location" id="location" value=""{% if errorMessage %} aria-describedby="error-message" aria-invalid="true"{% endif %}>
           </div>
