+++ config/routes.js
@@ -1,11 +1,10 @@
 const router = require('express').Router();
+const locationValidator = require('../app/middleware/locationValidator');
+const notInEnglandHandler = require('../app/middleware/notInEnglandHandler');
+const postcodeLookup = require('../app/middleware/postcodeLookup');
 const renderer = require('../app/middleware/renderer');
 const selectionValidatorSymptoms = require('../app/middleware/selectionValidatorSymptoms');
 const selectionValidatorAge = require('../app/middleware/selectionValidatorAge');
-const setLocals = require('../app/middleware/setLocals');
-const locationValidator = require('../app/middleware/locationValidator');
-const postcodeLookup = require('../app/middleware/postcodeLookup');
-const notInEnglandHandler = require('../app/middleware/notInEnglandHandler');
 
 router.get(
   '/',
@@ -14,39 +13,33 @@ router.get(
 
 router.get(
   '/symptoms',
-  setLocals.fromRequest,
   renderer.symptoms
 );
 
 router.get(
   '/recommend',
-  setLocals.fromRequest,
   selectionValidatorSymptoms,
   renderer.recommend
 );
 
 router.get(
   '/age',
-  setLocals.fromRequest,
   renderer.age
 );
 
 router.get(
   '/choose',
-  setLocals.fromRequest,
   selectionValidatorAge,
   renderer.choose
 );
 
 router.get(
   '/location',
-  setLocals.fromRequest,
   renderer.location
 );
 
 router.get(
   '/results',
-  setLocals.fromRequest,
   locationValidator,
   postcodeLookup,
   notInEnglandHandler,
